+++ src/components/justification-letter-pdf.tsx
@@ -3,20 +3,8 @@
 import React, { useState, useEffect, useRef } from "react";
 import { jsPDF } from 'jspdf';
 
-export default function JustificationLetterPDF() {
-    const [formData, setFormData] = useState(
-        {
-            managerName: '',
-            company: '',
-            specificProject: '',
-            specificTeam: '',
-            travelCost: '',
-            agenda: '',
-            goal: '',
-            name: '',
-            title: '',
-        }
-    )
+export default function JustificationLetterPDF(formData: any, setFormData: any) {
+
     const [isSmallScreen, setIsSmallScreen] = useState(false)
     const pdfRef = useRef<HTMLDivElement>(null);
     const currentDate = new Date();
@@ -30,13 +18,7 @@ export default function JustificationLetterPDF() {
         return () => window.removeEventListener("resize", handleResize)
     }, [])
 
-    const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const { name, value } = e.target;
-        setFormData(prevData => ({
-            ...prevData,
-            [name]: value
-        }))
-    }
+
 
     const handlePrint = () => {
         window.print();
@@ -83,91 +65,7 @@ export default function JustificationLetterPDF() {
     }
 
     return (
-        <div className="min-h-screen bg-gray-900 text-white md:flex md:gap-2">
-
-            <form className="rounded-lg border text-card-foreground shadow-sm bg-white border-gray-700 max-w-lg md:max-w-md mx-auto mt-8 p-8">
-                <label className="block text-sm font-medium mb-1 text-black">Manager&apos;s Name</label>
-                <input
-                    name="managerName"
-                    value={formData.managerName}
-                    onChange={handleChange}
-                    placeholder="Enter your manager's name"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Company</label>
-                <input
-                    name="company"
-                    value={formData.company}
-                    onChange={handleChange}
-                    placeholder="Enter your company name"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Specific Project</label>
-                <input
-                    name="specificProject"
-                    value={formData.specificProject}
-                    onChange={handleChange}
-                    placeholder="Enter your specific project"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Specific Team/Goal</label>
-                <input
-                    name="specificTeam"
-                    value={formData.specificTeam}
-                    onChange={handleChange}
-                    placeholder="Enter your specific team/goal"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Travel Cost</label>
-                <input
-                    name="travelCost"
-                    value={formData.travelCost}
-                    onChange={handleChange}
-                    placeholder="Enter your travel cost"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Agenda</label>
-                <input
-                    name="agenda"
-                    value={formData.agenda}
-                    onChange={handleChange}
-                    placeholder="Enter the conference agenda"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Goal</label>
-                <input
-                    name="goal"
-                    value={formData.goal}
-                    onChange={handleChange}
-                    placeholder="specific goal, e.g., modernize IT infrastructure, enhance cloud security, improve operational efficiency"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="text-sm font-medium mb-1 text-black">Your Name</label>
-                <input
-                    name="name"
-                    value={formData.name}
-                    onChange={handleChange}
-                    placeholder="Enter your name"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-                <label className="block text-sm font-medium mb-1 text-black">Your Title</label>
-                <input
-                    name="title"
-                    value={formData.title}
-                    onChange={handleChange}
-                    placeholder="Enter your title"
-                    className="w-full rounded border text-card-foreground shadow-sm bg-white border-gray-300 p-2 text-black mb-4"
-                />
-
-            </form >
+        <div className="min-h-screen bg-gray-900 text-white ">
 
             <section ref={pdfRef} className="rounded-lg border text-card-foreground shadow-sm bg-gray-800 border-gray-700 max-w-xl mx-auto mt-8 p-8 flex-grow">
                 <div className={`pdf-content ${isSmallScreen ? "text-sm" : "text-base"}`}>
